feat: discord verification
Add verification module responsible for managing the verifications
channel, updating the instructions, handling the interactions, and
managing access to the verification channel, exposing this to the
external modules.
Add database support with sqlx and mysql driver.
Split types package into types for each individual module, decoupling
the modules from each other.
Add localized strings for user interactions on verification.
Make DOOC messages require a verified account. If user is not verified
and verification is enabled in the config, user will be redirected to
the verification channel. Otherwise their discord username will be used.